Module google_analyticsadmin1_alpha::api

Structs

AccountDeleteCall
Marks target Account as soft-deleted (ie: "trashed") and returns it. This API does not have a method to restore soft-deleted accounts. However, they can be restored using the Trash Can UI. If the accounts are not restored before the expiration time, the account and all child resources (eg: Properties, GoogleAdsLinks, Streams, UserLinks) will be permanently purged. https://support.google.com/analytics/answer/6154772 Returns an error if the target is not found.

AccountGetCall
Lookup for a single Account.

AccountGetDataSharingSettingCall
Get data sharing settings on an account. Data sharing settings are singletons.

AccountListCall
Returns all accounts accessible by the caller. Note that these accounts might not currently have GA4 properties. Soft-deleted (ie: "trashed") accounts are excluded by default. Returns an empty list if no relevant accounts are found.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the GoogleAnalyticsAdmin hub.

AccountPatchCall
Updates an account.

AccountProvisionAccountTicketCall
Requests a ticket for creating an account.

AccountSearchChangeHistoryEventCall
Searches through all changes to an account or its children given the specified set of filters.

AccountSummaryListCall
Returns summaries of all accounts accessible by the caller.

AccountSummaryMethods
A builder providing access to all methods supported on accountSummary resources. It is not used directly, but through the GoogleAnalyticsAdmin hub.

AccountUserLinkAuditCall
Lists all user links on an account or property, including implicit ones that come from effective permissions granted by groups or organization admin roles. If a returned user link does not have direct permissions, they cannot be removed from the account or property directly with the DeleteUserLink command. They have to be removed from the group/etc that gives them permissions, which is currently only usable/discoverable in the GA or GMP UIs.

AccountUserLinkBatchCreateCall
Creates information about multiple users' links to an account or property. This method is transactional. If any UserLink cannot be created, none of the UserLinks will be created.

AccountUserLinkBatchDeleteCall
Deletes information about multiple users' links to an account or property.

AccountUserLinkBatchGetCall
Gets information about multiple users' links to an account or property.

AccountUserLinkBatchUpdateCall
Updates information about multiple users' links to an account or property.

AccountUserLinkCreateCall
Creates a user link on an account or property. If the user with the specified email already has permissions on the account or property, then the user's existing permissions will be unioned with the permissions specified in the new UserLink.

AccountUserLinkDeleteCall
Deletes a user link on an account or property.

AccountUserLinkGetCall
Gets information about a user's link to an account or property.

AccountUserLinkListCall
Lists all user links on an account or property.

AccountUserLinkPatchCall
Updates a user link on an account or property.

GoogleAnalyticsAdmin
Central instance to access all GoogleAnalyticsAdmin related resource activities

GoogleAnalyticsAdminV1alphaAccount
A resource message representing a Google Analytics account.

GoogleAnalyticsAdminV1alphaAccountSummary
A virtual resource representing an overview of an account and all its child GA4 properties.

GoogleAnalyticsAdminV1alphaAndroidAppDataStream
A resource message representing a Google Analytics Android app stream.

GoogleAnalyticsAdminV1alphaAuditUserLink
Read-only resource used to summarize a principal's effective roles.

GoogleAnalyticsAdminV1alphaAuditUserLinksRequest
Request message for AuditUserLinks RPC.

GoogleAnalyticsAdminV1alphaAuditUserLinksResponse
Response message for AuditUserLinks RPC.

GoogleAnalyticsAdminV1alphaBatchCreateUserLinksRequest
Request message for BatchCreateUserLinks RPC.

GoogleAnalyticsAdminV1alphaBatchCreateUserLinksResponse
Response message for BatchCreateUserLinks RPC.

GoogleAnalyticsAdminV1alphaBatchDeleteUserLinksRequest
Request message for BatchDeleteUserLinks RPC.

GoogleAnalyticsAdminV1alphaBatchGetUserLinksResponse
Response message for BatchGetUserLinks RPC.

GoogleAnalyticsAdminV1alphaBatchUpdateUserLinksRequest
Request message for BatchUpdateUserLinks RPC.

GoogleAnalyticsAdminV1alphaBatchUpdateUserLinksResponse
Response message for BatchUpdateUserLinks RPC.

GoogleAnalyticsAdminV1alphaChangeHistoryChange
A description of a change to a single Google Analytics resource.

GoogleAnalyticsAdminV1alphaChangeHistoryChangeChangeHistoryResource
A snapshot of a resource as before or after the result of a change in change history.

GoogleAnalyticsAdminV1alphaChangeHistoryEvent
A set of changes within a Google Analytics account or its child properties that resulted from the same cause. Common causes would be updates made in the Google Analytics UI, changes from customer support, or automatic Google Analytics system changes.

GoogleAnalyticsAdminV1alphaCreateUserLinkRequest
Request message for CreateUserLink RPC. Users can have multiple email addresses associated with their Google account, and one of these email addresses is the "primary" email address. Any of the email addresses associated with a Google account may be used for a new UserLink, but the returned UserLink will always contain the "primary" email address. As a result, the input and output email address for this request may differ.

GoogleAnalyticsAdminV1alphaDataSharingSettings
A resource message representing data sharing settings of a Google Analytics account.

GoogleAnalyticsAdminV1alphaDeleteUserLinkRequest
Request message for DeleteUserLink RPC.

GoogleAnalyticsAdminV1alphaEnhancedMeasurementSettings
Singleton resource under a WebDataStream, configuring measurement of additional site interactions and content.

GoogleAnalyticsAdminV1alphaFirebaseLink
A link between an GA4 property and a Firebase project.

GoogleAnalyticsAdminV1alphaGlobalSiteTag
Read-only resource with the tag for sending data from a website to a WebDataStream.

GoogleAnalyticsAdminV1alphaGoogleAdsLink
A link between an GA4 property and a Google Ads account.

GoogleAnalyticsAdminV1alphaIosAppDataStream
A resource message representing a Google Analytics IOS app stream.

GoogleAnalyticsAdminV1alphaListAccountSummariesResponse
Response message for ListAccountSummaries RPC.

GoogleAnalyticsAdminV1alphaListAccountsResponse
Request message for ListAccounts RPC.

GoogleAnalyticsAdminV1alphaListAndroidAppDataStreamsResponse
Request message for ListAndroidDataStreams RPC.

GoogleAnalyticsAdminV1alphaListFirebaseLinksResponse
Response message for ListFirebaseLinks RPC

GoogleAnalyticsAdminV1alphaListGoogleAdsLinksResponse
Response message for ListGoogleAdsLinks RPC.

GoogleAnalyticsAdminV1alphaListIosAppDataStreamsResponse
Request message for ListIosAppDataStreams RPC.

GoogleAnalyticsAdminV1alphaListPropertiesResponse
Response message for ListProperties RPC.

GoogleAnalyticsAdminV1alphaListUserLinksResponse
Response message for ListUserLinks RPC.

GoogleAnalyticsAdminV1alphaListWebDataStreamsResponse
Request message for ListWebDataStreams RPC.

GoogleAnalyticsAdminV1alphaProperty
A resource message representing a Google Analytics GA4 property.

GoogleAnalyticsAdminV1alphaPropertySummary
A virtual resource representing metadata for an GA4 property.

GoogleAnalyticsAdminV1alphaProvisionAccountTicketRequest
Request message for ProvisionAccountTicket RPC.

GoogleAnalyticsAdminV1alphaProvisionAccountTicketResponse
Response message for ProvisionAccountTicket RPC.

GoogleAnalyticsAdminV1alphaSearchChangeHistoryEventsRequest
Request message for SearchChangeHistoryEvents RPC.

GoogleAnalyticsAdminV1alphaSearchChangeHistoryEventsResponse
Response message for SearchAccounts RPC.

GoogleAnalyticsAdminV1alphaUpdateUserLinkRequest
Request message for UpdateUserLink RPC.

GoogleAnalyticsAdminV1alphaUserLink
A resource message representing a user's permissions on an Account or Property resource.

GoogleAnalyticsAdminV1alphaWebDataStream
A resource message representing a Google Analytics web stream.

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

PropertyAndroidAppDataStreamCreateCall
Creates an Android app stream with the specified location and attributes. Note that an Android app stream must be linked to a Firebase app to receive traffic. To create a working app stream, make sure your property is linked to a Firebase project. Then, use the Firebase API to create a Firebase app, which will also create an appropriate data stream in Analytics (may take up to 24 hours).

PropertyAndroidAppDataStreamDeleteCall
Deletes an android app stream on a property.

PropertyAndroidAppDataStreamGetCall
Lookup for a single AndroidAppDataStream

PropertyAndroidAppDataStreamListCall
Returns child android app streams under the specified parent property. Android app streams will be excluded if the caller does not have access. Returns an empty list if no relevant android app streams are found.

PropertyAndroidAppDataStreamPatchCall
Updates an android app stream on a property.

PropertyCreateCall
Creates an "GA4" property with the specified location and attributes.

PropertyDeleteCall
Marks target Property as soft-deleted (ie: "trashed") and returns it. This API does not have a method to restore soft-deleted properties. However, they can be restored using the Trash Can UI. If the properties are not restored before the expiration time, the Property and all child resources (eg: GoogleAdsLinks, Streams, UserLinks) will be permanently purged. https://support.google.com/analytics/answer/6154772 Returns an error if the target is not found, or is not an GA4 Property.

PropertyFirebaseLinkCreateCall
Creates a FirebaseLink. Properties can have at most one FirebaseLink.

PropertyFirebaseLinkDeleteCall
Deletes a FirebaseLink on a property

PropertyFirebaseLinkListCall
Lists FirebaseLinks on a property. Properties can have at most one FirebaseLink.

PropertyFirebaseLinkPatchCall
Updates a FirebaseLink on a property

PropertyGetCall
Lookup for a single "GA4" Property.

PropertyGoogleAdsLinkCreateCall
Creates a GoogleAdsLink.

PropertyGoogleAdsLinkDeleteCall
Deletes a GoogleAdsLink on a property

PropertyGoogleAdsLinkListCall
Lists GoogleAdsLinks on a property.

PropertyGoogleAdsLinkPatchCall
Updates a GoogleAdsLink on a property

PropertyIosAppDataStreamCreateCall
Creates an iOS app stream with the specified location and attributes. Note that an iOS app stream must be linked to a Firebase app to receive traffic. To create a working app stream, make sure your property is linked to a Firebase project. Then, use the Firebase API to create a Firebase app, which will also create an appropriate data stream in Analytics (may take up to 24 hours).

PropertyIosAppDataStreamDeleteCall
Deletes an iOS app stream on a property.

PropertyIosAppDataStreamGetCall
Lookup for a single IosAppDataStream

PropertyIosAppDataStreamListCall
Returns child iOS app data streams under the specified parent property. iOS app data streams will be excluded if the caller does not have access. Returns an empty list if no relevant iOS app data streams are found.

PropertyIosAppDataStreamPatchCall
Updates an iOS app stream on a property.

PropertyListCall
Returns child Properties under the specified parent Account. Only "GA4" properties will be returned. Properties will be excluded if the caller does not have access. Soft-deleted (ie: "trashed") properties are excluded by default. Returns an empty list if no relevant properties are found.

PropertyMethods
A builder providing access to all methods supported on property resources. It is not used directly, but through the GoogleAnalyticsAdmin hub.

PropertyPatchCall
Updates a property.

PropertyUserLinkAuditCall
Lists all user links on an account or property, including implicit ones that come from effective permissions granted by groups or organization admin roles. If a returned user link does not have direct permissions, they cannot be removed from the account or property directly with the DeleteUserLink command. They have to be removed from the group/etc that gives them permissions, which is currently only usable/discoverable in the GA or GMP UIs.

PropertyUserLinkBatchCreateCall
Creates information about multiple users' links to an account or property. This method is transactional. If any UserLink cannot be created, none of the UserLinks will be created.

PropertyUserLinkBatchDeleteCall
Deletes information about multiple users' links to an account or property.

PropertyUserLinkBatchGetCall
Gets information about multiple users' links to an account or property.

PropertyUserLinkBatchUpdateCall
Updates information about multiple users' links to an account or property.

PropertyUserLinkCreateCall
Creates a user link on an account or property. If the user with the specified email already has permissions on the account or property, then the user's existing permissions will be unioned with the permissions specified in the new UserLink.

PropertyUserLinkDeleteCall
Deletes a user link on an account or property.

PropertyUserLinkGetCall
Gets information about a user's link to an account or property.

PropertyUserLinkListCall
Lists all user links on an account or property.

PropertyUserLinkPatchCall
Updates a user link on an account or property.

PropertyWebDataStreamCreateCall
Creates a web stream with the specified location and attributes.

PropertyWebDataStreamDeleteCall
Deletes a web stream on a property.

PropertyWebDataStreamGetCall
Lookup for a single WebDataStream

PropertyWebDataStreamGetEnhancedMeasurementSettingCall
Returns the singleton enhanced measurement settings for this web stream. Note that the stream must enable enhanced measurement for these settings to take effect.

PropertyWebDataStreamGetGlobalSiteTagCall
Returns the Site Tag for the specified web stream. Site Tags are immutable singletons.

PropertyWebDataStreamListCall
Returns child web data streams under the specified parent property. Web data streams will be excluded if the caller does not have access. Returns an empty list if no relevant web data streams are found.

PropertyWebDataStreamPatchCall
Updates a web stream on a property.

PropertyWebDataStreamUpdateEnhancedMeasurementSettingCall
Updates the singleton enhanced measurement settings for this web stream. Note that the stream must enable enhanced measurement for these settings to take effect.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.