Struct google_analyticsadmin1_alpha::api::AccountMethods

pub struct AccountMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on account resources. It is not used directly, but through the GoogleAnalyticsAdmin hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_analyticsadmin1_alpha as analyticsadmin1_alpha;
 
use std::default::Default;
use oauth2;
use analyticsadmin1_alpha::GoogleAnalyticsAdmin;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = GoogleAnalyticsAdmin::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `get_data_sharing_settings(...)`, `list(...)`, `patch(...)`, `provision_account_ticket(...)`, `search_change_history_events(...)`, `user_links_audit(...)`, `user_links_batch_create(...)`, `user_links_batch_delete(...)`, `user_links_batch_get(...)`, `user_links_batch_update(...)`, `user_links_create(...)`, `user_links_delete(...)`, `user_links_get(...)`, `user_links_list(...)` and `user_links_patch(...)`
// to build up your call.
let rb = hub.accounts();

Implementations

impl<'a, C> AccountMethods<'a, C>

pub fn user_links_audit(
&self,
request: GoogleAnalyticsAdminV1alphaAuditUserLinksRequest,
parent: &str
) -> AccountUserLinkAuditCall<'a, C>

Create a builder to help you perform the following task:

Lists all user links on an account or property, including implicit ones that come from effective permissions granted by groups or organization admin roles. If a returned user link does not have direct permissions, they cannot be removed from the account or property directly with the DeleteUserLink command. They have to be removed from the group/etc that gives them permissions, which is currently only usable/discoverable in the GA or GMP UIs.

Arguments

request - No description provided.

parent - Required. Example format: accounts/1234

pub fn user_links_batch_create(
&self,
request: GoogleAnalyticsAdminV1alphaBatchCreateUserLinksRequest,
parent: &str
) -> AccountUserLinkBatchCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates information about multiple users' links to an account or property. This method is transactional. If any UserLink cannot be created, none of the UserLinks will be created.

Arguments

request - No description provided.

parent - Required. The account or property that all user links in the request are for. This field is required. The parent field in the CreateUserLinkRequest messages must either be empty or match this field. Example format: accounts/1234

pub fn user_links_batch_delete(
&self,
request: GoogleAnalyticsAdminV1alphaBatchDeleteUserLinksRequest,
parent: &str
) -> AccountUserLinkBatchDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes information about multiple users' links to an account or property.

Arguments

request - No description provided.

parent - Required. The account or property that all user links in the request are for. The parent of all values for user link names to delete must match this field. Example format: accounts/1234

pub fn user_links_batch_get(
&self,
parent: &str
) -> AccountUserLinkBatchGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about multiple users' links to an account or property.

Arguments

parent - Required. The account or property that all user links in the request are for. The parent of all provided values for the 'names' field must match this field. Example format: accounts/1234

pub fn user_links_batch_update(
&self,
request: GoogleAnalyticsAdminV1alphaBatchUpdateUserLinksRequest,
parent: &str
) -> AccountUserLinkBatchUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates information about multiple users' links to an account or property.

Arguments

request - No description provided.

parent - Required. The account or property that all user links in the request are for. The parent field in the UpdateUserLinkRequest messages must either be empty or match this field. Example format: accounts/1234

pub fn user_links_create(
&self,
request: GoogleAnalyticsAdminV1alphaUserLink,
parent: &str
) -> AccountUserLinkCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a user link on an account or property. If the user with the specified email already has permissions on the account or property, then the user's existing permissions will be unioned with the permissions specified in the new UserLink.

Arguments

request - No description provided.

parent - Required. Example format: accounts/1234

pub fn user_links_delete(&self, name: &str) -> AccountUserLinkDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a user link on an account or property.

Arguments

name - Required. Example format: accounts/1234/userLinks/5678

pub fn user_links_get(&self, name: &str) -> AccountUserLinkGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a user's link to an account or property.

Arguments

name - Required. Example format: accounts/1234/userLinks/5678

pub fn user_links_list(&self, parent: &str) -> AccountUserLinkListCall<'a, C>

Create a builder to help you perform the following task:

Lists all user links on an account or property.

Arguments

parent - Required. Example format: accounts/1234

pub fn user_links_patch(
&self,
request: GoogleAnalyticsAdminV1alphaUserLink,
name: &str
) -> AccountUserLinkPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a user link on an account or property.

Arguments

request - No description provided.

name - Example format: properties/1234/userLinks/5678

pub fn delete(&self, name: &str) -> AccountDeleteCall<'a, C>

Create a builder to help you perform the following task:

Marks target Account as soft-deleted (ie: "trashed") and returns it. This API does not have a method to restore soft-deleted accounts. However, they can be restored using the Trash Can UI. If the accounts are not restored before the expiration time, the account and all child resources (eg: Properties, GoogleAdsLinks, Streams, UserLinks) will be permanently purged. https://support.google.com/analytics/answer/6154772 Returns an error if the target is not found.

Arguments

name - Required. The name of the Account to soft-delete. Format: accounts/{account} Example: "accounts/100"

pub fn get(&self, name: &str) -> AccountGetCall<'a, C>

Create a builder to help you perform the following task:

Lookup for a single Account.

Arguments

name - Required. The name of the account to lookup. Format: accounts/{account} Example: "accounts/100"

pub fn get_data_sharing_settings(
&self,
name: &str
) -> AccountGetDataSharingSettingCall<'a, C>

Create a builder to help you perform the following task:

Get data sharing settings on an account. Data sharing settings are singletons.

Arguments

name - Required. The name of the settings to lookup. Format: accounts/{account}/dataSharingSettings Example: "accounts/1000/dataSharingSettings"

pub fn list(&self) -> AccountListCall<'a, C>

Create a builder to help you perform the following task:

Returns all accounts accessible by the caller. Note that these accounts might not currently have GA4 properties. Soft-deleted (ie: "trashed") accounts are excluded by default. Returns an empty list if no relevant accounts are found.

pub fn patch(
&self,
request: GoogleAnalyticsAdminV1alphaAccount,
name: &str
) -> AccountPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an account.

Arguments

request - No description provided.

name - Output only. Resource name of this account. Format: accounts/{account} Example: "accounts/100"

pub fn provision_account_ticket(
&self,
request: GoogleAnalyticsAdminV1alphaProvisionAccountTicketRequest
) -> AccountProvisionAccountTicketCall<'a, C>

Create a builder to help you perform the following task:

Requests a ticket for creating an account.

Arguments

request - No description provided.

pub fn search_change_history_events(
&self,
request: GoogleAnalyticsAdminV1alphaSearchChangeHistoryEventsRequest,
account: &str
) -> AccountSearchChangeHistoryEventCall<'a, C>

Create a builder to help you perform the following task:

Searches through all changes to an account or its children given the specified set of filters.

Arguments

request - No description provided.

account - Required. The account resource for which to return change history resources.