Struct google_analyticsadmin1_alpha::api::PropertyMethods

pub struct PropertyMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on property resources. It is not used directly, but through the GoogleAnalyticsAdmin hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_analyticsadmin1_alpha as analyticsadmin1_alpha;
 
use std::default::Default;
use oauth2;
use analyticsadmin1_alpha::GoogleAnalyticsAdmin;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = GoogleAnalyticsAdmin::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `android_app_data_streams_create(...)`, `android_app_data_streams_delete(...)`, `android_app_data_streams_get(...)`, `android_app_data_streams_list(...)`, `android_app_data_streams_patch(...)`, `create(...)`, `delete(...)`, `firebase_links_create(...)`, `firebase_links_delete(...)`, `firebase_links_list(...)`, `firebase_links_patch(...)`, `get(...)`, `google_ads_links_create(...)`, `google_ads_links_delete(...)`, `google_ads_links_list(...)`, `google_ads_links_patch(...)`, `ios_app_data_streams_create(...)`, `ios_app_data_streams_delete(...)`, `ios_app_data_streams_get(...)`, `ios_app_data_streams_list(...)`, `ios_app_data_streams_patch(...)`, `list(...)`, `patch(...)`, `user_links_audit(...)`, `user_links_batch_create(...)`, `user_links_batch_delete(...)`, `user_links_batch_get(...)`, `user_links_batch_update(...)`, `user_links_create(...)`, `user_links_delete(...)`, `user_links_get(...)`, `user_links_list(...)`, `user_links_patch(...)`, `web_data_streams_create(...)`, `web_data_streams_delete(...)`, `web_data_streams_get(...)`, `web_data_streams_get_enhanced_measurement_settings(...)`, `web_data_streams_get_global_site_tag(...)`, `web_data_streams_list(...)`, `web_data_streams_patch(...)` and `web_data_streams_update_enhanced_measurement_settings(...)`
// to build up your call.
let rb = hub.properties();

Implementations

impl<'a, C> PropertyMethods<'a, C>

pub fn android_app_data_streams_create(
&self,
request: GoogleAnalyticsAdminV1alphaAndroidAppDataStream,
parent: &str
) -> PropertyAndroidAppDataStreamCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an Android app stream with the specified location and attributes. Note that an Android app stream must be linked to a Firebase app to receive traffic. To create a working app stream, make sure your property is linked to a Firebase project. Then, use the Firebase API to create a Firebase app, which will also create an appropriate data stream in Analytics (may take up to 24 hours).

Arguments

request - No description provided.

parent - Required. The parent resource where this android app data stream will be created. Format: properties/123

pub fn android_app_data_streams_delete(
&self,
name: &str
) -> PropertyAndroidAppDataStreamDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an android app stream on a property.

Arguments

name - Required. The name of the android app data stream to delete. Format: properties/{property_id}/androidAppDataStreams/{stream_id} Example: "properties/123/androidAppDataStreams/456"

pub fn android_app_data_streams_get(
&self,
name: &str
) -> PropertyAndroidAppDataStreamGetCall<'a, C>

Create a builder to help you perform the following task:

Lookup for a single AndroidAppDataStream

Arguments

name - Required. The name of the android app data stream to lookup. Format: properties/{property_id}/androidAppDataStreams/{stream_id} Example: "properties/123/androidAppDataStreams/456"

pub fn android_app_data_streams_list(
&self,
parent: &str
) -> PropertyAndroidAppDataStreamListCall<'a, C>

Create a builder to help you perform the following task:

Returns child android app streams under the specified parent property. Android app streams will be excluded if the caller does not have access. Returns an empty list if no relevant android app streams are found.

Arguments

parent - Required. The name of the parent property. For example, to limit results to app streams under the property with Id 123: "properties/123"

pub fn android_app_data_streams_patch(
&self,
request: GoogleAnalyticsAdminV1alphaAndroidAppDataStream,
name: &str
) -> PropertyAndroidAppDataStreamPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an android app stream on a property.

Arguments

request - No description provided.

name - Output only. Resource name of this Data Stream. Format: properties/{property_id}/androidAppDataStreams/{stream_id} Example: "properties/1000/androidAppDataStreams/2000"

pub fn firebase_links_create(
&self,
request: GoogleAnalyticsAdminV1alphaFirebaseLink,
parent: &str
) -> PropertyFirebaseLinkCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a FirebaseLink. Properties can have at most one FirebaseLink.

Arguments

request - No description provided.

parent - Required. Format: properties/{property_id} Example: properties/1234

pub fn firebase_links_delete(
&self,
name: &str
) -> PropertyFirebaseLinkDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a FirebaseLink on a property

Arguments

name - Required. Format: properties/{property_id}/firebaseLinks/{firebase_link_id} Example: properties/1234/firebaseLinks/5678

pub fn firebase_links_list(
&self,
parent: &str
) -> PropertyFirebaseLinkListCall<'a, C>

Create a builder to help you perform the following task:

Lists FirebaseLinks on a property. Properties can have at most one FirebaseLink.

Arguments

parent - Required. Format: properties/{property_id} Example: properties/1234

pub fn firebase_links_patch(
&self,
request: GoogleAnalyticsAdminV1alphaFirebaseLink,
name: &str
) -> PropertyFirebaseLinkPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a FirebaseLink on a property

Arguments

request - No description provided.

name - Output only. Example format: properties/1234/firebaseLinks/5678

pub fn google_ads_links_create(
&self,
request: GoogleAnalyticsAdminV1alphaGoogleAdsLink,
parent: &str
) -> PropertyGoogleAdsLinkCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a GoogleAdsLink.

Arguments

request - No description provided.

parent - Required. Example format: properties/1234

pub fn google_ads_links_delete(
&self,
name: &str
) -> PropertyGoogleAdsLinkDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a GoogleAdsLink on a property

Arguments

name - Required. Example format: properties/1234/googleAdsLinks/5678

pub fn google_ads_links_list(
&self,
parent: &str
) -> PropertyGoogleAdsLinkListCall<'a, C>

Create a builder to help you perform the following task:

Lists GoogleAdsLinks on a property.

Arguments

parent - Required. Example format: properties/1234

pub fn google_ads_links_patch(
&self,
request: GoogleAnalyticsAdminV1alphaGoogleAdsLink,
name: &str
) -> PropertyGoogleAdsLinkPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a GoogleAdsLink on a property

Arguments

request - No description provided.

name - Output only. Format: properties/{propertyId}/googleAdsLinks/{googleAdsLinkId} Note: googleAdsLinkId is not the Google Ads customer ID.

pub fn ios_app_data_streams_create(
&self,
request: GoogleAnalyticsAdminV1alphaIosAppDataStream,
parent: &str
) -> PropertyIosAppDataStreamCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an iOS app stream with the specified location and attributes. Note that an iOS app stream must be linked to a Firebase app to receive traffic. To create a working app stream, make sure your property is linked to a Firebase project. Then, use the Firebase API to create a Firebase app, which will also create an appropriate data stream in Analytics (may take up to 24 hours).

Arguments

request - No description provided.

parent - Required. The parent resource where this ios app data stream will be created. Format: properties/123

pub fn ios_app_data_streams_delete(
&self,
name: &str
) -> PropertyIosAppDataStreamDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an iOS app stream on a property.

Arguments

name - Required. The name of the iOS app data stream to delete. Format: properties/{property_id}/iosAppDataStreams/{stream_id} Example: "properties/123/iosAppDataStreams/456"

pub fn ios_app_data_streams_get(
&self,
name: &str
) -> PropertyIosAppDataStreamGetCall<'a, C>

Create a builder to help you perform the following task:

Lookup for a single IosAppDataStream

Arguments

name - Required. The name of the iOS app data stream to lookup. Format: properties/{property_id}/iosAppDataStreams/{stream_id} Example: "properties/123/iosAppDataStreams/456"

pub fn ios_app_data_streams_list(
&self,
parent: &str
) -> PropertyIosAppDataStreamListCall<'a, C>

Create a builder to help you perform the following task:

Returns child iOS app data streams under the specified parent property. iOS app data streams will be excluded if the caller does not have access. Returns an empty list if no relevant iOS app data streams are found.

Arguments

parent - Required. The name of the parent property. For example, to list results of app streams under the property with Id 123: "properties/123"

pub fn ios_app_data_streams_patch(
&self,
request: GoogleAnalyticsAdminV1alphaIosAppDataStream,
name: &str
) -> PropertyIosAppDataStreamPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an iOS app stream on a property.

Arguments

request - No description provided.

name - Output only. Resource name of this Data Stream. Format: properties/{property_id}/iosAppDataStreams/{stream_id} Example: "properties/1000/iosAppDataStreams/2000"

pub fn user_links_audit(
&self,
request: GoogleAnalyticsAdminV1alphaAuditUserLinksRequest,
parent: &str
) -> PropertyUserLinkAuditCall<'a, C>

Create a builder to help you perform the following task:

Lists all user links on an account or property, including implicit ones that come from effective permissions granted by groups or organization admin roles. If a returned user link does not have direct permissions, they cannot be removed from the account or property directly with the DeleteUserLink command. They have to be removed from the group/etc that gives them permissions, which is currently only usable/discoverable in the GA or GMP UIs.

Arguments

request - No description provided.

parent - Required. Example format: accounts/1234

pub fn user_links_batch_create(
&self,
request: GoogleAnalyticsAdminV1alphaBatchCreateUserLinksRequest,
parent: &str
) -> PropertyUserLinkBatchCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates information about multiple users' links to an account or property. This method is transactional. If any UserLink cannot be created, none of the UserLinks will be created.

Arguments

request - No description provided.

parent - Required. The account or property that all user links in the request are for. This field is required. The parent field in the CreateUserLinkRequest messages must either be empty or match this field. Example format: accounts/1234

pub fn user_links_batch_delete(
&self,
request: GoogleAnalyticsAdminV1alphaBatchDeleteUserLinksRequest,
parent: &str
) -> PropertyUserLinkBatchDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes information about multiple users' links to an account or property.

Arguments

request - No description provided.

parent - Required. The account or property that all user links in the request are for. The parent of all values for user link names to delete must match this field. Example format: accounts/1234

pub fn user_links_batch_get(
&self,
parent: &str
) -> PropertyUserLinkBatchGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about multiple users' links to an account or property.

Arguments

parent - Required. The account or property that all user links in the request are for. The parent of all provided values for the 'names' field must match this field. Example format: accounts/1234

pub fn user_links_batch_update(
&self,
request: GoogleAnalyticsAdminV1alphaBatchUpdateUserLinksRequest,
parent: &str
) -> PropertyUserLinkBatchUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates information about multiple users' links to an account or property.

Arguments

request - No description provided.

parent - Required. The account or property that all user links in the request are for. The parent field in the UpdateUserLinkRequest messages must either be empty or match this field. Example format: accounts/1234

pub fn user_links_create(
&self,
request: GoogleAnalyticsAdminV1alphaUserLink,
parent: &str
) -> PropertyUserLinkCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a user link on an account or property. If the user with the specified email already has permissions on the account or property, then the user's existing permissions will be unioned with the permissions specified in the new UserLink.

Arguments

request - No description provided.

parent - Required. Example format: accounts/1234

pub fn user_links_delete(&self, name: &str) -> PropertyUserLinkDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a user link on an account or property.

Arguments

name - Required. Example format: accounts/1234/userLinks/5678

pub fn user_links_get(&self, name: &str) -> PropertyUserLinkGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a user's link to an account or property.

Arguments

name - Required. Example format: accounts/1234/userLinks/5678

pub fn user_links_list(&self, parent: &str) -> PropertyUserLinkListCall<'a, C>

Create a builder to help you perform the following task:

Lists all user links on an account or property.

Arguments

parent - Required. Example format: accounts/1234

pub fn user_links_patch(
&self,
request: GoogleAnalyticsAdminV1alphaUserLink,
name: &str
) -> PropertyUserLinkPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a user link on an account or property.

Arguments

request - No description provided.

name - Example format: properties/1234/userLinks/5678

pub fn web_data_streams_create(
&self,
request: GoogleAnalyticsAdminV1alphaWebDataStream,
parent: &str
) -> PropertyWebDataStreamCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a web stream with the specified location and attributes.

Arguments

request - No description provided.

parent - Required. The parent resource where this web data stream will be created. Format: properties/123

pub fn web_data_streams_delete(
&self,
name: &str
) -> PropertyWebDataStreamDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a web stream on a property.

Arguments

name - Required. The name of the web data stream to delete. Format: properties/{property_id}/webDataStreams/{stream_id} Example: "properties/123/webDataStreams/456"

pub fn web_data_streams_get(
&self,
name: &str
) -> PropertyWebDataStreamGetCall<'a, C>

Create a builder to help you perform the following task:

Lookup for a single WebDataStream

Arguments

name - Required. The name of the web data stream to lookup. Format: properties/{property_id}/webDataStreams/{stream_id} Example: "properties/123/webDataStreams/456"

pub fn web_data_streams_get_enhanced_measurement_settings(
&self,
name: &str
) -> PropertyWebDataStreamGetEnhancedMeasurementSettingCall<'a, C>

Create a builder to help you perform the following task:

Returns the singleton enhanced measurement settings for this web stream. Note that the stream must enable enhanced measurement for these settings to take effect.

Arguments

name - Required. The name of the settings to lookup. Format: properties/{property_id}/webDataStreams/{stream_id}/enhancedMeasurementSettings Example: "properties/1000/webDataStreams/2000/enhancedMeasurementSettings"

pub fn web_data_streams_get_global_site_tag(
&self,
name: &str
) -> PropertyWebDataStreamGetGlobalSiteTagCall<'a, C>

Create a builder to help you perform the following task:

Returns the Site Tag for the specified web stream. Site Tags are immutable singletons.

Arguments

name - Required. The name of the site tag to lookup. Note that site tags are singletons and do not have unique IDs. Format: properties/{property_id}/webDataStreams/{stream_id}/globalSiteTag Example: "properties/123/webDataStreams/456/globalSiteTag"

pub fn web_data_streams_list(
&self,
parent: &str
) -> PropertyWebDataStreamListCall<'a, C>

Create a builder to help you perform the following task:

Returns child web data streams under the specified parent property. Web data streams will be excluded if the caller does not have access. Returns an empty list if no relevant web data streams are found.

Arguments

parent - Required. The name of the parent property. For example, to list results of web streams under the property with Id 123: "properties/123"

pub fn web_data_streams_patch(
&self,
request: GoogleAnalyticsAdminV1alphaWebDataStream,
name: &str
) -> PropertyWebDataStreamPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a web stream on a property.

Arguments

request - No description provided.

name - Output only. Resource name of this Data Stream. Format: properties/{property_id}/webDataStreams/{stream_id} Example: "properties/1000/webDataStreams/2000"

pub fn web_data_streams_update_enhanced_measurement_settings(
&self,
request: GoogleAnalyticsAdminV1alphaEnhancedMeasurementSettings,
name: &str
) -> PropertyWebDataStreamUpdateEnhancedMeasurementSettingCall<'a, C>

Create a builder to help you perform the following task:

Updates the singleton enhanced measurement settings for this web stream. Note that the stream must enable enhanced measurement for these settings to take effect.

Arguments

request - No description provided.

name - Output only. Resource name of this Data Stream. Format: properties/{property_id}/webDataStreams/{stream_id}/enhancedMeasurementSettings Example: "properties/1000/webDataStreams/2000/enhancedMeasurementSettings"

pub fn create(
&self,
request: GoogleAnalyticsAdminV1alphaProperty
) -> PropertyCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an "GA4" property with the specified location and attributes.

Arguments

request - No description provided.

pub fn delete(&self, name: &str) -> PropertyDeleteCall<'a, C>

Create a builder to help you perform the following task:

Marks target Property as soft-deleted (ie: "trashed") and returns it. This API does not have a method to restore soft-deleted properties. However, they can be restored using the Trash Can UI. If the properties are not restored before the expiration time, the Property and all child resources (eg: GoogleAdsLinks, Streams, UserLinks) will be permanently purged. https://support.google.com/analytics/answer/6154772 Returns an error if the target is not found, or is not an GA4 Property.

Arguments

name - Required. The name of the Property to soft-delete. Format: properties/{property_id} Example: "properties/1000"

pub fn get(&self, name: &str) -> PropertyGetCall<'a, C>

Create a builder to help you perform the following task:

Lookup for a single "GA4" Property.

Arguments

name - Required. The name of the property to lookup. Format: properties/{property_id} Example: "properties/1000"

pub fn list(&self) -> PropertyListCall<'a, C>

Create a builder to help you perform the following task:

Returns child Properties under the specified parent Account. Only "GA4" properties will be returned. Properties will be excluded if the caller does not have access. Soft-deleted (ie: "trashed") properties are excluded by default. Returns an empty list if no relevant properties are found.

pub fn patch(
&self,
request: GoogleAnalyticsAdminV1alphaProperty,
name: &str
) -> PropertyPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a property.

Arguments

request - No description provided.

name - Output only. Resource name of this property. Format: properties/{property_id} Example: "properties/1000"